Ensure that erase cannot split the blockchain into 2 disconnected parts (#20)
* GenericBlockRepository: implement isInUse()

* BitcoinStore: add a check that erasing a block must not split the blockchain into 2 disconnected parts

* VeriBlockStore: add a check that erasing a block must not split the blockchain into 2 disconnected parts

* GenericBlockRepository: clean up isInUse()

+++ reference-lib/src/main/java/org/veriblock/integrations/blockchain/store/BitcoinStore.java
@@ -98,6 +98,11 @@ public class BitcoinStore {
 
     public StoredBitcoinBlock erase(Sha256Hash hash) throws BlockStoreException, SQLException {
         StoredBitcoinBlock erased = get(hash);
+
+        if(erased != null && bitcoinRepository.isInUse(hash)) {
+            throw new BlockStoreException("Cannot erase a block referenced by another block");
+        }
+
         bitcoinRepository.delete(hash);
         return erased;
     }
